Starting implementing pawn promotion and better memory efficiency for moves
This started as a simple implementation of the pawn promotion, i noticed that the overridden method for ExecuteMove on pawnpromotion was not called because i used an array of Moves and not an array of pointers. Now it's not working, maybe I nedd to use smartpointers instead of normal pointers.